requirementsfinal.txt
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.12.10
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
businesstimedelta==1.0.1
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
convertdate==2.4.0
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
Django==4.1.1
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.2
django-braces==1.15.0
django-compressor==4.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==2.9.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
excel2json==1.0.1
excel2json-3==0.1.6
exponent-server-sdk==1.0.3
hijri-converter==2.2.4
holidays==0.16
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
isort==5.10.1
kombu==5.2.4
korean-lunar-calendar==0.3.1
lazy-object-proxy==1.7.1
MarkupPy==1.14
mccabe==0.7.0
numpy==1.23.3
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pylint==2.15.3
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pylint-quotes==0.2.3
PyMeeus==0.5.11
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-excel2json==0.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
rollbar==0.16.3
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
tomli==2.0.1
tomlkit==0.11.4
Twisted==22.8.0
twython==3.9.1
txaio==22.2.1
types-cryptography==3.3.23
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
